From now on I'll make the log format easier to view.
1.Recontest the training contest of bjtu.
p1001 - Clear All of Them I
DP. using 2^20 to save the state. when come to a state i, using lowbit to get its first point doesn't get matched. Enum the other point to match with it and update the former state if it has a optimizing result. The answer of the problem must be the f[2^20-1].

p1002 - Clear All of Them II
Hasn't solved in the contest. Simulation and DP. It's easier than the p1001. Every time use the spfa to get the valid four points and update the answer.

p1004 - Invoker
Polya thorem.

p1005 - Extended-Tree(Unsolved)
construct a three-children tree in the i-th sequence. Divide and conquer.

p1006 - Substring
given two integer a(<= 10^100) and b(<= 10^7), check the minimum addition for a to contain a sunstring of b.
if a <= b, the answer is obviously b-a.
otherwise, enum the position of the occurence in a of b(fron len(b) to len(a)+1, if b=0, stop at len(a)).
for each enum, the compare b with the next len(b) digits start from position i:
if it's larger than b, then add 1 to the position in a higher than i by 1. set the digits after substring b in a to 0.
otherwise if it's smaller than b, simply set the digits after substring b in a to 0.
if it's equals b, then output the answer 0 imediatlly.

p1007 - Hand in hand
simple problem. because the vertice has no more than two edge, the gragh only can made up by some circle and chain. if the type of chain and circle in two gragh are same, it's valid.

p1008 - Math Geek
There's a expression to construct the anti magic square. But in this problem, the random algorithm to swap two position is enough.

2.Two problems in same type of game theory
Beijing 2008 Regional/UVA Liva-Archive 4323 - a simple stone game
Firstly, we should check out if a position is a P-position. We start out in solve out serveral cases for k=5.
	for all the k<=5 it's must be the P-position.
	p=6	-> P
	p=7	-> N
	p=8	-> P
	p=9	-> N
	p=10	-> P
	p=11	-> N
	p=12	-> P
	p=13,14 -> N
	p=15	-> P
	p=16,17	-> N
	p=18	-> P
	p=19,20,21	-> N
	p=22	-> P
	p=23,24,25,26	-> N
	p=27	-> P
	p=28,29,30,31,32	-> N
	p=33	-> P
	p=34,35,36,37,38,39,40(2+),41 -> N; p=40(1) -> P
	p=42	-> P
	p=43,44,45,46,47,48(2+),49,50(2+),51	-> N; p=48(1),50(1)	-> P
	k=52	-> P
	p=53,54,55,56,57,58(2+),59,60(2+),61,62(2+),63	-> N; p=58(1),60(1),62(1)	-> P
	p=64	-> P
	p=65,66,67,68,69,70(2+),71,72(2+),73,74(2+),75,76(3+),77,78,79	-> N; p=70(1),72(1),74(1),76(2)	-> P

From the data above that, we can infer that:
1). If a postion can go to the P postion in the safe step(the next player cannot using previous steps to get to the terminal), the position is a N-position.
2). Additionally, we can find there're several positions that will become P-postion if the previous step is equal or smaller than a certain number. Because though it can reach the P postion in safe steps, but there're not enough steps to reach it, and it cannot reach another postions of this kind. We defininate this kind of postion small P-position. note that small P-position is not a lose state for the frist player. Moreover, besides the postions that can reach the P-postion within one safe steps, some positions can firstly reach the small P-postion and gradually get the P-postion. See p=41 as an example. It cannot reach 33 within safe step, but it can go to 40 first, because 40(1) is a small P-position.

If we can find all the N-postion and small P-postions after a P postion, the next P postion must just after them. The N-postion of first type is easy to determine, but how to get the postions for the type two. We can transfer the problem to such child problem: we must get to the P-postion at last, then we see P-postion's new index is 0, and we're now at n-P-postion. We can step forward for (P-postion-1)/k steps for the first move to ensure the next player cannot step to the real terminal. At this point, the problem f(n) is divided to the problem f(n-P-postion). As the result, what's the next P-postion after the current P-postion? we can find that the postion NXT - P-postion must be exactly the index of P-postion after (P-postion-1)/k. Because it's the first postion that after the (P-postion-1)/k and cannot go to the last small P-position within (P-postion-1)/k. As the result, p[i+1] = p[i]+p[next[i]]. next[i] is the P index of the first P-postion that bigger than (p[i]-1)/k.

We have already infered that the lose state can only happen at P-postion other than small P-postion, so the first problem has solved.

Secondly, if it's not the P-postion, how can we find the smallest step?
It's become easy when we get the conclusion above. We can divide f(n) to f(n-(P-postion)) easily noted by f(n1). If n1 is not the P-postion, it's obvious that it's not the small P-postion in the last function(f(n)), then it can be divided further(to n2, n3, ...). If it comes to the condition that ni is P-postion, it must be a small P-position for the f(n[i-1]). So it must step forward to ni steps, to reach its nearest small P-postion.(If n1 is P-position, it's a small P-position initially, and must step to the really P-position only). The problem then solved.

-------------END-----------------

HDU2516 - Stone game
A more eaier game in the same theory with the problem above, where k is always 2 and only need to check if the first player will win.