1.CodeForces Round 95 Division 2 1).proA - cAPS lOCK A mixed meaning problem of implementation. If the format of given string is in the certain condition, transform the whole string. 2).proB - Opposites Attract Hash problem. find the number of pair that v1 = -v2. A trick: 0 cannot be the partner of itself. 3).proC - The World is a Theatre Recurrence problem. let f[i] be the number of condition that boys is exactly i(4 <= i < t). ignore the condition of i > n and t-i > m. 4).proD - Yet Another Task with Queens The pair of queens who are the neighbour in the same line/column/diagonal line are threaten each other. it canbe exactly found by using sort in line/column/diagonal line order. Total time is O(nlogn).
2.CodeForces Round 96 Division 1
1).proA - Turning Tape
Another mixed meaning problem of implementation. Reverse the previous letter in binary format as the key(0 as the result of first calculation), minuse the reverse of the current letter. result will be added by 256 if it's negative.
2).proC - Logo Turtle
A simple dp problem. Let f[i][j][k] to be the i-th step using j times reversation, and the orintation is k(0 for backward and 1 for forward). f[i][j][k] can contribute to (f[i+1][j][k] and f[i+1][j+1][1-k]) or (f[i+1][j][1-k] and f[i+1][j+1][k]). The answer is the best one for f[n][j][k], where m-j is even.
dp for two times. first time start at f[0][0][0] = 0. second time start at f[0][0][1] = 0.
3).proD - Constants in the language of Shakespeare
Greedy algorithm.
the first time combine the consecutive '1' together(if i..j is 1, we can using +2^(j+1) and -2^i to archive the goal).
the second time we combine the two neighbour consecutive unit the previous step get if the distance of the tail of previous and the head of latter is 1 and not both are a single 1.(if i..j and j+2..k is in the condition, we can using +2^(k+1), -2^(j+1) and -2^i to archive the goal). The combined unit in this step can be combined to other unit again.
then we count the rest unit, if it's only constructed by a single 1, just count it by one. else count it by the number of tiniest units added by 1. and the way to output it is easy(told before).
My fault in the contest: I firstly didn't combine the unit at following condition in step 2: the distance of two units is 1, but one of the unit is constructed by a single 1. The truth is, combine the unit won't let the answer worse, and may let it optimized. examples below:
111010111 -> combined will let it optimized.
11101 -> combined won't let it be worse.

3.Computational Geometry in POJ
POJ3432 - Count Squares
Find the number of squares that the vertexes is exactly on the given point. Using hash map and enum the edge of square. O(N^2*logN).
POJ3512 - Incidental Points
Find a line wich go across maximum number of given points. enum each point and find the optimized set which has maximum number of points who are in the same orientation to the base point. the answer is the number of largest set. Be careful with condition that multi points in the same position. O(N^2*logN).

4.TopCoder SRM 525 Division 1
proA - DropCoins
A implementation.Given a grid and some coins in the integer point, and the left, right, upper, low bound of the rectangle. for all the sub-rectangle, which has the property of the distance of the left bound and the root's left bound dl, the distance between the right bound and the root's right bound, the distance between the upper bound and root's upper bound du, the distance between the low bound and the root's low bound, the number of coins in the sub-rectangle dk. find a sub-rectangle whose dk equals given k and let the (min(dl,dr)+min(du,dl))*2+(max(dl,dr)+max(du,dl)) be minimum.
proB - Rumor
Enum the strategy of each rabit when it first get a rumor(0 or 1). If all the rabits who lastly get the first rumor can get both rumor, return the number of the day. Otherwise output the number of the day + 1.
My fault in contest: I firstly only enum the strategy of the start vertex(who initially know the rumor), and the wrong case is as follow:
1->3->
       5->6->7
2->4->
1,2 is the start vertex. The answer is 6(day+1), but only count the strategy of start point can get the answer of 5(day).