two problems about Number Theory: 1) POJ 2635 - The Embarrassed Cryptographer tag: prime sift, high-precision module description: given a large number k (k<=10^100), and it's guaranteed that k is the product of two prime. The goal is to check whether the smaller prime is smaller than another given number L(L<=10^6). solution: sift the prime list first. then use k module all the prime which is smaller than L. note that the large number k should be stored in at least thousand-decimal to avoid the TLE.
2)POJ 3292 - Semi-prime H-numbers
tag: prime sift
description: find out all the semi-prime h-numbers decripted in the problem.
solution: we can see the h-number as the normal number, then we find the prime. In the prime sift process, once we find a prime, we times the prime and others we found before. Obviously, they're all semi-prime h-numbers. Then we get the answer.