POJ 3294 - Life Forms (Using Linear Sufix Array)
There is a fatal problem when calculating answer: I put a tag to check if the next bundle is valid to put into answer set. The tag is refresh when ans is bigger or the monotonous queue is empty, but the latter thing may happen when the next element is the same as the smallest one in queue, it may cause program output same string many times(generate 1000x100 data can check it out). Put the special judge at this point will get ac.
status:
1. OLE for many times.
2. at last AC.