HDU 3060 - Area 2
Problem - Calculate the area of Union of two simple polygon
Solution - The same idea as polyxcircle. The difference is: now it's two triangle instead of circle and triangle. The two triangles share a same vertex origin, and the orientation area is positive iff there cross product has the same sign.
Attention - For the first loop, if the two point's cross product is zero, there is no need to calculation deeper for this two.
Result - The problem has the wrong test data, see the discuss.