POJ 2195 - Going Home
Problem : Raw Minum cost Maximum flow.
Description: First time to complete the MCMF. Confirmed that dinic is not suitable for MCMF, since it will fall to dead loop when there is a zero-loop (if without too much more code). At the same time, it won't loose the time consume like maximum flow. Thus EK is enough. This time the SPFA structure is nearly completed, however, the EK structure is still need to optimize.