HDU 1506 - Largest Rectangle In a Histogram
Problem : Given a histogram, find a largest area rectangle in it.
Solution : RAW monotone stack problem. Denote f[i] to be largest rectangle passes i and with height a[i]. f can be calculated in O(n) time by using monotone stack.
Note : The property of monotone stack is: if the element on top of the stack is not ended calculating, so does the elements below it.